+++ src/rust/src/ios/ios_platform.rs
@@ -109,14 +109,10 @@ impl Platform for IosPlatform {
             audio_levels_interval,
         );
 
-        let audio_jitter_buffer_max_packets: i32 = call_config
-            .audio_jitter_buffer_max_packets
-            .try_into()
-            .expect("isize fits in an i32");
-        let audio_jitter_buffer_max_target_delay_ms: i32 = call_config
-            .audio_jitter_buffer_max_target_delay_ms
-            .try_into()
-            .expect("isize fits in an i32");
+        let audio_jitter_buffer_max_packets = call_config.audio_jitter_buffer_config.max_packets;
+        let audio_jitter_buffer_max_target_delay_ms =
+            call_config.audio_jitter_buffer_config.max_target_delay_ms;
+
         let connection = Connection::new(
             call.clone(),
             remote_device_id,
@@ -414,8 +410,6 @@ impl Platform for IosPlatform {
         message: Vec<u8>,
         urgency: group_call::SignalingMessageUrgency,
     ) -> Result<()> {
-        info!("send_call_message():");
-
         (self.app_interface.sendCallMessage)(
             self.app_interface.object,
             app_slice_from_bytes(Some(&recipient_uuid)),